feat: ErrorCode 추가
