+++ frontend/techpick/src/apis/apiClient.ts
@@ -1,7 +1,8 @@
 import * as Sentry from '@sentry/nextjs';
 import ky, { HTTPError } from 'ky';
+import { ERROR_MESSAGE_JSON } from '@/constants';
 import { notifyError } from '@/utils';
-import { ERROR_MESSAGE_JSON, returnErrorFromHTTPError } from './error';
+import { returnErrorFromHTTPError } from './error';
 
 export const apiClient = ky.create({
   credentials: 'include',
@@ -24,11 +25,6 @@ export const apiClient = ky.create({
             /* empty */
           } else if (ERROR_MESSAGE_JSON[errorCode]) {
             notifyError(ERROR_MESSAGE_JSON[errorCode]);
-          } else if (errorCode === 'PK-000') {
-            // TODO: 익스텐션에서 가져왔습니다. PK-000은 다른 기능에서도 작동하기 때문에 수정해야 합니다.
-            /* empty */
-          } else if (errorCode === 'UNKNOWN') {
-            notifyError('서버에서 알 수 없는 에러가 발생했습니다.');
           }
         }
