+++ frontend/techpick/src/constants/errorMessageJson.ts
@@ -1,0 +1,24 @@
+export const ERROR_MESSAGE_JSON: { [key: string]: string } = {
+  'PK-000': '존재하지 않는 픽입니다.',
+  'PK-001': '이미 존재하는 픽입니다! 픽 저장을 실패했습니다!',
+  'PK-002': '접근할 수 없는 픽입니다.',
+  'PK-003': '접근할 수 없는 픽입니다.',
+  'PK-004': '픽은 휴지통에서만 삭제할 수 있습니다.',
+  'PK-005': '제목이 허용된 최대길이를 초과했습니다.',
+  'FO-000': '존재하지 않는 폴더입니다.',
+  'FO-001': '이미 존재하는 폴더 이름입니다.',
+  'FO-002': '접근할 수 없는 폴더입니다.',
+  'FO-007': '이동하려는 폴더의 범위가 올바르지 않습니다.',
+  'SF-000': '존재하지 않는 공유 폴더입니다. 접근할 수 없습니다.',
+  'SF-001': '접근할 수 없는 공유 폴더입니다.',
+  'SF-002': '해당 폴더는 공유할 수 없습니다.',
+  'SF-003': '이미 공유된 폴더는 다시 공유할 수 없습니다.',
+  'TG-000': '존재하지 않는 태그입니다.',
+  'TG-001': '중복된 태그가 존재합니다.',
+  'TG-002': '유효하지 않는 태그 이름입니다.',
+  'TG-003': '잘못된 태그에 접근했습니다.',
+  'LI-000': '존재하지 않는 링크입니다.',
+  'LI-002': '이미 존재하는 링크입니다.',
+  'LI-003': '이 링크는 넣을 수 없습니다. ㅠㅠ',
+  UNKNOWN: '죄송합니다. 서버에서 알 수 없는 에러가 발생했습니다.',
+};
