+++ frontend/techpick/src/constants/index.ts
@@ -2,3 +2,4 @@ export const PORTAL_CONTAINER_ID = 'portalContainer';
 export const UNKNOWN_FOLDER_ID = -9999;
 export { ROUTES } from './route';
 export { COLOR_LIST } from './colorList';
+export { ERROR_MESSAGE_JSON } from './errorMessageJson';
